+++ previous-versions/accounting/src/sdk/models/shared/createbillpaymentresponse.ts
@@ -266,7 +266,7 @@ export class CreateBillPaymentResponse extends SpeakeasyBase {
     requestedOnUtc: string;
 
     /**
-     * The status of the push operation.
+     * The current status of the push operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })