+++ previous-versions/accounting/src/sdk/models/shared/createdirectcostresponse.ts
@@ -130,7 +130,7 @@ export class CreateDirectCostResponse extends SpeakeasyBase {
     requestedOnUtc: string;
 
     /**
-     * The status of the push operation.
+     * The current status of the push operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })