+++ previous-versions/accounting/src/sdk/models/shared/createdirectincomeresponse.ts
@@ -131,7 +131,7 @@ export class CreateDirectIncomeResponse extends SpeakeasyBase {
     requestedOnUtc: string;
 
     /**
-     * The status of the push operation.
+     * The current status of the push operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })