+++ previous-versions/accounting/src/sdk/models/shared/createitemresponse.ts
@@ -121,7 +121,7 @@ export class CreateItemResponse extends SpeakeasyBase {
     requestedOnUtc: string;
 
     /**
-     * The status of the push operation.
+     * The current status of the push operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })