+++ previous-versions/accounting/src/sdk/models/shared/createjournalentryresponse.ts
@@ -137,7 +137,7 @@ export class CreateJournalEntryResponse extends SpeakeasyBase {
     requestedOnUtc: string;
 
     /**
-     * The status of the push operation.
+     * The current status of the push operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })