+++ previous-versions/accounting/src/sdk/models/shared/createtransferresponse.ts
@@ -118,7 +118,7 @@ export class CreateTransferResponse extends SpeakeasyBase {
     requestedOnUtc: string;
 
     /**
-     * The status of the push operation.
+     * The current status of the push operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })