+++ previous-versions/accounting/src/sdk/models/shared/updatebankaccountresponse.ts
@@ -113,7 +113,7 @@ export class UpdateBankAccountResponse extends SpeakeasyBase {
     requestedOnUtc: string;
 
     /**
-     * The status of the push operation.
+     * The current status of the push operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })