+++ previous-versions/accounting/RELEASES.md
@@ -679,3 +679,13 @@ Based on:
 - [typescript v0.37.0] previous-versions/accounting
 ### Releases
 - [NPM v0.37.0] https://www.npmjs.com/package/@codat/accounting/v/0.37.0 - previous-versions/accounting
+
+## 2023-09-21 00:03:51
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.88.1 (2.122.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.37.1] previous-versions/accounting
+### Releases
+- [NPM v0.37.1] https://www.npmjs.com/package/@codat/accounting/v/0.37.1 - previous-versions/accounting